% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ES}
\alias{ES}
\title{Calculates GSEA-like enrichment scores.}
\usage{
ES(p, w, pws, both = FALSE)
}
\arguments{
\item{p}{The name of the pathway for which the enrichment score is to be
calculated.}

\item{w}{A ranked vector of values/weights.}

\item{pws}{A vector with as many entries as in w, assigning a pathway to each
entry in w.}

\item{both}{Optional logical parameter. Should the enrichment score be
separated by negative/positive values.}
}
\value{
The enrichment score.
}
\description{
Calculates GSEA-like enrichment scores.
}
\examples{
 NULL

}

