% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exarray.R
\name{eset_reduce}
\alias{eset_reduce}
\title{Reduces an ExpressionSet by an n-to-n map of features to groups.}
\usage{
eset_reduce(eset, features, groups, method = "mean", progress = TRUE)
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{features}{A character vector of features to be grouped.}

\item{groups}{A factor or character vector mapping the entries in
\code{features} to groups.}

\item{method}{The reduction method. Must be one of 'mean', 'median' or 'max'.
Note that these are the geometric versions, meaning they operate on the log
expression values.}

\item{progress}{Should progress information be shown.}
}
\value{
A new ExpressionSet with features given by \code{unique(groups)}.
}
\description{
All entries in \code{features} must exist in \code{eset}. \code{features} and
\code{groups} must have the same length. Note that the 'mean' and 'median'
are geometric variants meaning they operate on the log-expression values
rather than on the raw expression values which usually gives better results.
}

