% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getemail.R
\name{getDevEmail}
\alias{getDevEmail}
\title{get maintainer email for a single package}
\usage{
getDevEmail(x, ...)
}
\arguments{
\item{x}{character(1) package name}

\item{\dots}{passed to biocBuildEmail, excluding textOnly, dry.run; version = "release" or "devel" will be important}
}
\description{
get maintainer email for a single package
}
