% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getemail.R
\name{getLandingPage}
\alias{getLandingPage}
\title{make URL for landing page}
\usage{
getLandingPage(x, version = "release")
}
\arguments{
\item{x}{character(1) package name}

\item{version}{character(1) "release" or "devel"}
}
\description{
make URL for landing page
}
