% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getemail.R
\name{getRangeTable}
\alias{getRangeTable}
\title{make table of package names and email addresses for a set of packages}
\usage{
getRangeTable(
  start,
  end,
  condition = "ERROR",
  phase = "buildsrc",
  version = "release",
  dump = TRUE
)
}
\arguments{
\item{start}{character(1) package name beginning range}

\item{end}{character(1) package name end of range}

\item{condition}{character(1) "ERROR" or "WARNINGS",}

\item{phase}{character(1) "buildsrc", "checksrc", "install", "buildbin"}

\item{version}{character(1) "release" (default) or "devel"}

\item{dump}{logical(1) if TRUE (default) produce pastable text}
}
\description{
make table of package names and email addresses for a set of packages
}
\examples{
getRangeTable("ReactomePA", "SLqPCR")
}
