% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{inter}
\alias{inter}
\title{Calculates interaction terms for a gene expression matrix.}
\usage{
inter(M)
}
\arguments{
\item{M}{A gene expression matrix with genes in the columns and samples
in the rows.}
}
\value{
A new gene expression matrix with n + 0.5*n*(n-1) columns, where n
 is the number of columns in the matrix M.
}
\description{
Interaction terms for two expression values g[i] and g[j] are given
by g[i]*g[j].
}
\examples{
 NULL

}

