% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{measures}
\alias{measures}
\title{Various measures for goodness-of-fit.}
\usage{
measures(truth, pred)
}
\arguments{
\item{truth}{A numeric vector describing the true values.}

\item{pred}{A numeric vector describing the predicted values.}
}
\value{
A numeric vector of length 5 containing the following metrics:
\describe{
 \item{mse}{The mean squared error.}
 \item{rmse}{The root mean squared error.}
 \item{mae}{The mean absolute error.}
 \item{mre}{The mean relative error.}
 \item{rsq}{The r-squared value.}
}
}
\description{
Calculates a set of five different goodness-of-fit metrics.
}
\examples{
 NULL

}

