% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{NES}
\alias{NES}
\title{Calculates the normalized enrichment scores.}
\usage{
NES(p, w, pws, n = 200)
}
\arguments{
\item{p}{The name of the pathway for which the enrichment score is to be
calculated.}

\item{w}{A ranked vector of values/weights.}

\item{pws}{A vector with as many entries as in w, assigning a pathway to each
entry in w.}

\item{n}{The number of random permutations to generate.}
}
\value{
A vector with two elements: the normalized enrichment score and an
empirical, uncorrected p-value for it.
}
\description{
Calculates es/mean(perm), where perm is a set of n random permutations
of pathway labels.
}
\examples{
 NULL

}

