% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tissue_lfc}
\alias{tissue_lfc}
\title{Calculates flux log-fold changes between in-tissue and out-tissue samples.}
\usage{
tissue_lfc(v, map, extra = NULL)
}
\arguments{
\item{v}{A matrix of fluxes where rows denote samples across different
tissues and columns denote fluxes.}

\item{map}{A map translating the row names of v to their specific tissue.}

\item{extra}{Optional data.frame with as many rows as rows in v that contains
additional information about the fluxes that should be appended to the results.}
}
\value{
A data.table containing the log-fold changes for each flux and tissue.
}
\description{
`tissue_lfc` takes values for fluxes across different tissues and calculates
the log-fold changes between the flux in a given target tissue against all
all other tissues, for all tissues.
}
\examples{
 NULL

}

